+++ wgpu/src/api/instance.rs
@@ -17,6 +17,8 @@ bitflags::bitflags! {
         const UnrestrictedPointerParameters = 1 << 2;
         /// <https://www.w3.org/TR/WGSL/#language_extension-pointer_composite_access>
         const PointerCompositeAccess = 1 << 3;
+        /// <https://www.w3.org/TR/WGSL/#language_extension-immediate_address_space>
+        const ImmediateAddressSpace = 1 << 4;
     }
 }